basictest available to run

+++ src/main/java/com/example/Shippable.java
@@ -4,6 +4,6 @@ import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public interface Shippable {
-    double calculateShippingCost();
+    BigDecimal calculateShippingCost();
     double weight();
 }
